Partial AOI Implementation
- Ships can now join/exit zones and also inform other ships of their
entering/leaving
- Tested well with 1 ship, but breaks when testing with 2 ships It's
buggy. For some reason, if there are 2 ships in play {1, 2} and they
both occupy, say, zone 1. If ship 2 leaves zone 1 it won't update. If
ship 1 leaves zone 1 it updates, but ship 2 might still be registered as
in the zone even though it clearly left it.

It seems that if there are > 1 ship in a zone, only the first ship can
exit/leave properly? Not sure. Need to test some more.